+++ client/src/sugarcube-2/completions.ts
@@ -1,4 +1,10 @@
+<<<<<<< HEAD
 import * as vscode from 'vscode';
+=======
+import * as fs from 'fs';
+import * as vscode from 'vscode';
+import { URL } from 'url';
+>>>>>>> parent of acb069b... Revert "jsdoc branch"
 import { CompletionItemKind } from 'vscode-languageclient';
 
 interface jsdocParsed {
@@ -45,6 +51,7 @@ const createSnippetDocs = function (target: jsdocParsed) {
     return r;
 };
 
+<<<<<<< HEAD
 export const completion = function (context: vscode.ExtensionContext) {
     let r: Array<vscode.CompletionItem> = [];
     const docs: Array<jsdocParsed> = Object.values(context.workspaceState.get("jsdocs", {}));
@@ -61,5 +68,28 @@ export const completion = function (context: vscode.ExtensionContext) {
             r.push(snippet);
         });
     });
+=======
+export const completion = function () {
+    let r: Array<vscode.CompletionItem> = [];
+
+    const ws = vscode.workspace.workspaceFolders?.[0].uri.path;
+    if (ws) {
+        const wsURL = new URL("file://" + ws + "/.vscode/.t3lt_temp.json");
+        const docs: Array<jsdocParsed> = Object.values(JSON.parse(fs.readFileSync(wsURL, "utf8")));
+
+        docs.forEach((el: jsdocParsed) => {
+            const names = [el.name].concat(el.alias?.split(" ") || []);
+
+            names.forEach(elem => {
+                const snippet = new vscode.CompletionItem(elem, CompletionItemKind.Method);
+                snippet.insertText = new vscode.SnippetString(createSnippet(el));
+                snippet.documentation = new vscode.MarkdownString(createSnippetDocs(el));
+                snippet.detail = el.kind || "";
+
+                r.push(snippet);
+            });
+        });
+    }
+>>>>>>> parent of acb069b... Revert "jsdoc branch"
     return r;
 };