+++ CHANGELOG.md
@@ -1,20 +1,36 @@
 # Changelog
 
+## v0.2.0
+
+Adds:
+
+- Syntax highlighting for Chapbook 1.
+
+- JSDoc comment support for SugarCube. See [Readme](README.md#features) for more information.
+
+Fixes:
+
+- SugarCube comments (`/* */`) not getting tokenized.
+
+---
+
 ## v0.1.0
 
 Initial release.
 
-- Adds syntax highlighting for Twee 3.
+Adds:
 
-- Adds dynamic snippet for generating IFID for the `StoryData` special passage. Typing 'StoryData' provides a completion prompt which inserts a JSON chunk along with the generated IFID.
+- Syntax highlighting for Twee 3.
 
-- Adds the following diagnostics:
+- Dynamic snippet for generating IFID for the `StoryData` special passage. Typing 'StoryData' provides a completion prompt which inserts a JSON chunk along with the generated IFID.
+
+- The following diagnostics:
     - *`Error`* Malformed `StoryData` JSON,
     - *`Warning`* No whitespace between start token `::` and passage name.
 
-- Adds automatic detection of storyformat plus an override option.
+- Automatic detection of storyformat plus an override option.
 
-- Adds syntax highlighting for:
+- Syntax highlighting for:
     - SugarCube 2,
     - Harlowe 3.